chore: trim the LOD comments to what the annotated code guarantees
Review feedback: the comments narrated what other code does with the value
rather than the value itself. Shortened to the rationale, and the test's
64799 is now a named constant.

+++ Explorer/Assets/DCL/LOD/Tests/EditMode/UpdateSceneLODInfoSystemShould.cs
@@ -111,7 +111,7 @@ namespace DCL.LOD.Tests
 
             //Assert
             var sceneLODInfoRetrieved = world.Get<SceneLODInfo>(entity);
-            Assert.AreEqual(1, sceneLODInfoRetrieved.CurrentLODLevelPromise);
+            Assert.AreEqual((byte)1, sceneLODInfoRetrieved.CurrentLODLevelPromise);
             Assert.AreEqual($"{sceneId}_{DIGEST}_1{PlatformUtils.GetCurrentPlatform()}", sceneLODInfoRetrieved.CurrentLODPromise.LoadingIntention.Hash);
         }
     }
